examples/code_llama/requirements.txt
llama-cpp-python
ray==2.8.0
ray[serve]==2.8.0